requirements.txt
streamlit==1.31.0
docling==2.41.0
requests==2.32.2
python-dotenv==1.0.0
chromadb
langchain
langchain_community
google-generativeai
nltk
PyPDF2
scikit-learn